requirements.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.1
certifi==2021.10.8
charset-normalizer==2.0.11
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-markdown==0.8.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
idna==3.3
importlib-metadata==4.10.1
itypes==1.2.0
Jinja2==3.0.3
kombu==4.6.11
Markdown==3.3.6
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2021.3
PyYAML==6.0
redis==3.4.1
requests==2.27.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
vine==1.3.0
zipp==3.7.0
